Renamed adc_common_v3 to adc_common_v1_multi
