+++ include/libopencm3/stm32/common/adc_common_v1_multi.h
